yarn -> npm, split sanitizer, update Actions
