+++ .github/workflows/build.yml
@@ -17,20 +17,16 @@ jobs:
       matrix:
         node-version: ['18.x']
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
       - name: Use Node.js ${{ matrix.node-version }}
-        uses: actions/setup-node@v3
+        uses: actions/setup-node@v4
         with:
           node-version: ${{ matrix.node-version }}
-      - name: Cache node_modules
-        uses: actions/cache@v3
-        with:
-          path: node_modules
-          key: modules-cache-v1-${{ runner.os }}-${{ hashFiles('yarn.lock') }}
-      - run: yarn install
-      - run: yarn lint
-      - run: yarn test
-      - run: yarn build
+          cache: npm
+      - run: npm ci
+      - run: npm run lint
+      - run: npm run test
+      - run: npm run build
       - name: save the built
         uses: actions/upload-artifact@v3
         if: "!failure()"
@@ -77,8 +73,8 @@ jobs:
       - build
     if: startsWith(github.ref, 'refs/tags/v')
     steps:
-      - uses: actions/checkout@v3
-      - uses: actions/setup-node@v3
+      - uses: actions/checkout@v4
+      - uses: actions/setup-node@v4
         with:
           tag_name: 'v%s'
           node-version: '18.x'
