+++ README.md
@@ -49,7 +49,7 @@ You can see more examples at [https://geolonia.github.io/embed/](https://geoloni
 ```shell
 $ git clone git@github.com:geolonia/embed.git
 $ cd embed
-$ yarn
+$ npm install
 $ npm start # run dev server
 $ npm test # run tests
 $ npm run build # build production bundle
