+++ package.json
@@ -30,6 +30,7 @@
     "@geolonia/eslint-config": "latest",
     "@types/jsdom": "^21.1.7",
     "@types/mocha": "^10.0.1",
+    "@types/sanitize-html": "^2.13.0",
     "babel-loader": "^9.1.0",
     "css-loader": "^6.7.2",
     "eslint": "^8.29.0",
