+++ tsconfig.json
@@ -6,8 +6,8 @@
       "dom.iterable",
       "esnext"
     ],
+    "outDir": "dist",
     "declaration": true,
-    "declarationDir": "dist",
     "skipLibCheck": true,
     "sourceMap": true,
     "esModuleInterop": true,
